+++ src/NHibernate/Cfg/NullConfigurationManager.cs
@@ -1,0 +1,20 @@
+namespace NHibernate.Cfg
+{
+	class NullConfigurationManager : ConfigurationProvider
+	{
+		public override IHibernateConfiguration GetConfiguration()
+		{
+			return null;
+		}
+
+		public override string GetNamedConnectionString(string name)
+		{
+			return null;
+		}
+
+		public override string GetLoggerFactoryClassName()
+		{
+			return null;
+		}
+	}
+}
